requirements.txt
flask == 1.1.1
flask-login
Flask-SQLAlchemy
flask-wtf
flask-script
pygal
newsapi-python
gunicorn
psycopg2